# perf: reduce copies on get/set_variable