Another ~12% improvement.
```
(kat) pelito@MacBook-Pro-7 katalyn % hyperfine -N -r 5 "./narivm_new ./bench.nvm" "./narivm_old ./bench.nvm"
Benchmark 1: ./narivm_new ./bench.nvm
  Time (mean ± σ):      7.352 s ±  0.324 s    [User: 6.858 s, System: 0.425 s]
  Range (min … max):    6.980 s …  7.654 s    5 runs
 
Benchmark 2: ./narivm_old ./bench.nvm
  Time (mean ± σ):      8.266 s ±  0.515 s    [User: 7.759 s, System: 0.364 s]
  Range (min … max):    7.779 s …  9.006 s    5 runs
 
Summary
  ./narivm_new ./bench.nvm ran
    1.12 ± 0.09 times faster than ./narivm_old ./bench.nvm
```